# docs/openapi.yml
openapi : "3.0.3"
info:
  title : Baan management API overview
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
paths:
...
  /bhaai/{bhaaiId}/baan:
    parameters:
      - $ref: '#/components/parameters/BhaaiId'
    get:
      security:
        - JWT: [ ]
      tags:
        - Baan
      operationId: getBaan
      summary: to get all baan in bhaai 
      responses:
        '200':
          description: List of baan
          content:
            application/json; chartset=utf-8:
              schema:
                $ref: '#/components/schemas/BaanList'
    post:
      security:
        - JWT: [ ]
      tags:
        - Baan
      operationId: createBaan 
      summary: to create baan
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/BaanBase'
      responses:
        "200":
          description: New created Bhaai object
          content:
            applocation/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/BaanList'
  /bhaai/{bhaaiId}/baan/{id}:
    parameters:
      - $ref: '#/components/parameters/BhaaiId'
      - $ref: '#/components/parameters/ResourceId'
    put:
      security:
        - JWT: [ ]
      tags:
        - Baan
      operationId: updateBaanById
      summary: to update baan by id
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/BaanBase'
      responses:
        '200':
          description:  Update baan
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Baan'   
    delete:
      security:
        - JWT: [ ]
      tags:
        - Baan
      operationId: deleteBaanById
      summary: to delete an by id 
      responses:
        '200':
          description: ok
  /search:
    parameters:
      - $ref: '#/components/parameters/SearchBy'
    get:
      security:
        - JWT: [ ]
      tags:
        - Baan
      operationId: searchBaan
      summary: to search baan 
      responses:
        '200':
          description:  Searched baan
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/BaanList'
...
  /pariwar:
    get:
      security:
        - JWT: [ ]
      tags:
        - Pariwar
      operationId: getListPariwar
      summary: to get all pariwar
      responses:
        '200':
          description: List of pariwar
          content:
            application/json; chartset=utf-8:
              schema:
                $ref: '#/components/schemas/PariwarList'
    post:
      security:
        - JWT: [ ]
      tags:
        - Pariwar
      operationId: createPariwar
      summary: to create pariwar
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref:  '#/components/schemas/PariwarBase'
      responses:
        "200":
          description: New created pariwar object
          content:
            applocation/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/PariwarList'
  /pariwar/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    put:
      security:
        - JWT: [ ]
      tags:
        - Pariwar
      operationId: updatePariwarId
      summary: to update pariwar
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/PariwarBase'
      responses:
        '200':
          description:  Update pariwar object
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/PariwarList'
    delete:
      security:
        - JWT: [ ]
      tags:
        - Pariwar
      operationId: deletePariwarById
      summary: to delete pariwar by id
      responses:
        '200':
          description: ok
...
  /pariwar/{pariwarId}/nimta:
    parameters:
      - $ref: '#/components/parameters/PariwarId'
    get:
      security:
        - JWT: [ ]
      tags:
        - Nimta
      operationId: getNimtaList
      summary: to get a list of all nimta
      responses:
        '200':
          description: Get nimta list
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/NimtaList'
    post:
      security:
        - JWT: [ ]
      tags:
        - Nimta
      operationId:  createNimta  
      summary:  to create a nimta
      requestBody:
        content:
          application/json; charset=utf-8:
            schema:
              $ref: '#/components/schemas/NimtaBase'
      responses:
        '200':
          description: create nimta
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/Nimta'
...
components:
  parameters:
    SearchBy:
      description: text for search
      in: query
      name: 'searchBy'
      schema:
        type: string
        example: godara64022@gmail.com
...
    PariwarId:
      required: true
      description: id of pariwar to be worked on 
      in: path 
      name: 'pariwarId'
      schema:
        type: string
        example: adasdas-dwqd3-fcsd-dwedd
    ResourceId:
      required: true
      description: Id of the resource to be worked on
      in: path
      name: 'id'
      schema:
        type: string
        example: adasdas-dwqd3-fcsd-dwedd
    BhaaiId:
      required: true
      description: Id of bhaai we worked oen
      in: path
      name: 'bhaaiId'
      schema:
        type: string
        example: adasdas-dwqd3-fcsd-dwedd
...
  schemas:
...
    PariwarBase:
      type: object
      properties:
        name:
          type: string
          example: shubham
      required:
        - name
      additionalProperties: false
...
    NimtaBase:
      type: object 
      additionalProperties: false
      properties:
        name:
          type: string
          example: shubham
        relatives:
          type: array
          items:
            type: string
            example: brother
      required:
        - name
        - relatives
...
    Pariwar:
      type: object
      additionalProperties: false
      allOf:
        - $ref: '#/components/schemas/PariwarBase'
      properties:
        _id:
          type: string
          example: aaas-ssss-dddd-dddd
      required:
        - _id
    Nimta:
      type: object
      additionalProperties: false
      allOf:
        - $ref: '#/components/schemas/NimtaBase'
      properties:
        _id:
          type: string
          example: aaas-ssss-dddd-dddd
      required:
        - _id
...
    PariwarList:
      type: array
      items:
        $ref: '#/components/schemas/Pariwar'
...
    NimtaList:
      type: array
      items: 
        $ref: '#/components/schemas/NimtaBase'
...
    BaanList:
      type: array
      items:
        $ref: '#/components/schemas/Baan'
    BaanBase:
      type: object
      properties:
        firstName:
          type: string
          example: shubham
        lastName:
          type: string
          example: godara  
        fathersName:
          type: string
          example: bhagwan dass
        address:
          type: string
          example: mp rohi
        nickName:
          type: string
          example: gholio 
        amount:
          type: number
          example: 12345 
      required:
        -  firstName
        -  lastName 
        -  fathersName
        -  address
        -  nickName
        -  amount
      additionalProperties: false 
    Baan:
      type: object
      allOf:
        - $ref: '#/components/schemas/BaanBase'
      properties:
        _id:
          type: string
          format: uuid
          example: fsdfsdf-fsdfsdf-sdfsdf-fsdf3
        BhaaiId:
          type: string
          format: uuid
          example: addf-dddfg-hhjj-hhjk
      required:
        - _id
        - bhaaiId
      additionalProperties: false
...
  securitySchemes:
    JWT:
      bearerFormat: JWT
      description: authantication
      scheme: bearer
      type: http